add Hamiltonian1D.py

+++ Anas/classes.py
@@ -118,7 +118,8 @@ class CNN:
 
         conv,dense,bias = params
 
-        x = sigma.reshape(1,1,self.L)
+        # Ensure input dtype matches conv kernel dtype for JAX convolution.
+        x = sigma.astype(conv.dtype).reshape(1,1,self.L)
 
         x = jax.lax.conv_general_dilated(
             x,
